requirements.txt
Flask==3.0.3
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
Flask-WTF==1.2.1
Flask-Limiter==3.7.0
Werkzeug==3.0.3
openpyxl==3.1.2
gunicorn==22.0.0
python-dotenv==1.0.1
